remove oss terms
