+++ ai/integrations/crewai/src/unitycatalog/ai/crewai/toolkit.py
@@ -5,7 +5,7 @@ from typing import Any, Callable, Dict, List, Optional
 from crewai_tools import BaseTool as CrewAIBaseTool
 from pydantic import BaseModel, ConfigDict, Field, model_validator
 
-from unitycatalog.ai.core.client import BaseFunctionClient
+from unitycatalog.ai.core.base import BaseFunctionClient
 from unitycatalog.ai.core.utils.client_utils import validate_or_set_default_client
 from unitycatalog.ai.core.utils.function_processing_utils import (
     generate_function_input_params_schema,
