+++ ai/integrations/openai/src/unitycatalog/ai/openai/utils.py
@@ -2,7 +2,7 @@ import json
 from typing import Any, Dict, List, Optional
 
 from openai.types.chat.chat_completion import ChatCompletion
-from unitycatalog.ai.core.client import BaseFunctionClient
+from unitycatalog.ai.core.base import BaseFunctionClient
 from unitycatalog.ai.core.utils.client_utils import validate_or_set_default_client
 from unitycatalog.ai.core.utils.function_processing_utils import construct_original_function_name
